WIP: Refactor and testing
